+++ packages/cli/package.json
@@ -32,6 +32,7 @@
     "@autorelease/jira": "6.5.0",
     "@autorelease/npm": "6.5.0",
     "@autorelease/released": "6.5.0",
+    "@autorelease/slack": "6.5.0",
     "chalk": "^2.4.2",
     "command-line-args": "^5.1.1",
     "command-line-usage": "^5.0.5",
