+++ plugins/slack/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*", "../../typings/**/*"]
+}
